motor driver implementations

+++ Chapter05/src/motor1.c
@@ -1,5 +1,5 @@
 
-int motor1_adjust_motor(int target_force){
+int motor1_adjust_motor(double target_force){
   //real driver would do something here
   //interface with physical components
   return 0;
