+++ Chapter05/src/motor2.c
@@ -1,5 +1,5 @@
 
-int motor2_adjust_motor(int target_force){
+int motor2_adjust_motor(double target_force){
   //real driver would do something here
   //interface with physical components
   return 0;
